requirements.txt
atlas-mpl-style==0.17.3
certifi==2020.12.5
chardet==4.0.0
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
matplotlib==3.3.4
mplhep==0.2.18
numexpr==2.7.3
numpy==1.20.1
packaging==20.9
palettable==3.3.0
Pillow==8.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
requests==2.25.1
scipy==1.6.1
six==1.15.0
uproot==4.0.6
urllib3==1.26.4
